Changed location of ekos product
